feat: use debounce on contrast checking
